Support decimal seriesNumber for interstitial/companion volumes
seriesNumber was Int?, which can't represent a real interstitial release
like "book 1.5" — the prior workaround was leaving those unnumbered
entirely. Widened to Float? and switched the remaining parseInt() call
sites (admin.js, bulk-import.js) to parseFloat().

Also fixes the local migration history: prisma/migrations/ was missing
the original initial-schema migration file (only a later fragment
existed), which made `prisma migrate dev` fail against a fresh shadow
database. Replaced that fragment with a baseline migration representing
the full schema as it stood before this change, recorded as already-
applied — the repo's migration folder is now self-consistent and can
migrate a fresh database from empty to current state.

+++ scripts/bulk-import.js
@@ -79,7 +79,7 @@ function normalizeRow(row) {
     year:         row.year ? parseInt(row.year) : null,
     author:       row.author ? String(row.author).trim() : null,
     seriesName:   row.seriesName ? String(row.seriesName).trim() : null,
-    seriesNumber: row.seriesNumber ? parseInt(row.seriesNumber) : null,
+    seriesNumber: row.seriesNumber ? parseFloat(row.seriesNumber) : null,
     tags: Array.isArray(row.tags)
       ? row.tags
       : (row.tags ? String(row.tags).split(';').map(t => t.trim()).filter(Boolean) : []),
